+++ lib/src/widgets/operator_icon.dart
@@ -2,20 +2,20 @@ import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'package:pars_validator/pars_validator.dart';
 
-class BankIcon extends StatelessWidget {
-  final Bank bank;
+class OperatorIcon extends StatelessWidget {
+  final MobileOperator operator;
   final double size;
 
-  const BankIcon({
+  const OperatorIcon({
     super.key,
-    required this.bank,
+    required this.operator,
     this.size = 24,
   });
 
   @override
   Widget build(BuildContext context) {
     return SvgPicture.asset(
-      'packages/pars_validator/assets/banks/icons/${bank.iconPath}',
+      'packages/pars_validator/assets/operators/icons/${operator.iconPath}',
       width: size,
       height: size,
       fit: BoxFit.fitWidth,